server_generator/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
anyio==4.9.0
APScheduler==3.10.4
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
Brotli==1.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
exceptiongroup==1.3.0
Flask==3.1.0
frozenlist==1.6.2
g4f==0.5.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.4.4
nest-asyncio==1.6.0
propcache==0.3.1
pycryptodome==3.23.0
python-dotenv==1.0.1
python-telegram-bot==20.7
pytz==2025.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
typing_extensions==4.14.0
tzlocal==5.3.1
urllib3==2.4.0
watchdog==6.0.0
Werkzeug==3.1.3
yarl==1.20.0
zipp==3.21.0
fastapi==0.109.2
uvicorn==0.27.1
g4f>=0.1.9.0
pydantic==2.6.1
starlette==0.36.3
typing-extensions>=4.8.0
anyio>=4.0.0
celery>=5.3.0
redis>=4.5.0
